requirements.txt
# Core
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
openai>=1.40.0

# Bioinformatics
scanpy>=1.10.0
pydeseq2>=0.4.0
anndata>=0.10.0
matplotlib>=3.8.0
seaborn>=0.13.0
adjustText>=1.0.0

# Infra
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0

# Testing
pytest>=8.0.0
